requirements.txt
asgiref==3.4.1
async-timeout==3.0.1
beautifulsoup4==4.9.3
bidict==0.21.2
click==8.0.1
Django==3.2.6
django-bootstrap4==3.0.1
dnspython==1.16.0
eventlet==0.30.2
greenlet==1.1.1
gunicorn==20.1.0
h11==0.12.0
python-engineio==4.2.1
python-socketio==5.4.0
pytz==2021.1
redis==3.5.3
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
whitenoise
django-redis
dj_database_url
psycopg2-binary